Upload Images and Deleted Images

+++ src/restaurants/restaurants.controller.ts
@@ -1,8 +1,9 @@
-import { Controller, Get, Post, Body, Param, Put, Delete, Query } from '@nestjs/common';
+import { Controller, Get, Post, Body, Param, Put, Delete, Query, UseInterceptors, UploadedFiles } from '@nestjs/common';
 import { Query as ExpressQuery} from 'express-serve-static-core';
 import { RestaurantsService } from './restaurants.service';
 import { Restaurant } from './schemas/restaurant.schema';
 import { CreateRestaurantDto, UpdateRestaurantDto } from './dtos/index';
+import { FilesInterceptor } from '@nestjs/platform-express';
 
 
 @Controller('restaurants')
@@ -33,8 +34,27 @@ export class RestaurantsController {
     };
 
     @Delete(':id')
-    async deleteRestaurant(@Param('id') id: string): Promise<Restaurant>{
-        await this.restauranstService.findById(id);
-        return this.restauranstService.deleteById(id)
+    async deleteRestaurant(@Param('id') id: string): Promise<{deleted: Boolean}>{
+        const restaurant = await this.restauranstService.findById(id);
+        const isDeleted = await this.restauranstService.deleteImages(restaurant.images);
+
+        if(isDeleted){
+            this.restauranstService.deleteById(id);
+
+            return {
+                deleted: true,
+            };
+        }else{
+            return {
+                deleted: false,
+            };
+        }
     };
+
+    @Put('upload/:id')
+    @UseInterceptors(FilesInterceptor('files'))
+    async uploadFiles(@Param('id') id:string, @UploadedFiles() files: Array<Express.Multer.File>){
+        await this.restauranstService.findById(id);
+        return await this.restauranstService.uploadImages(id,files);
+    }
 }
